Auto_Marks/automarks_BE/run_api.py
#!/usr/bin/env python3
"""
API Runner Script
Runs the FastAPI application for VTU Results Management System
"""
import os
import uvicorn
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

if __name__ == "__main__":
    uvicorn.run(
        "app.main:app",
        host=os.getenv("API_HOST", "0.0.0.0"),
        port=int(os.getenv("API_PORT", 8000)),
        reload=os.getenv("API_RELOAD", "True") == "True"
    )