Auto_Marks/automarks_BE/pyproject.toml
[project]
name = "vtu-results-system"
version = "0.1.0"
description = "VTU Results Extraction and Management System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.12",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "sqlalchemy==2.0.36",
    "psycopg2-binary==2.9.10",
    "alembic==1.14.0",
    "pandas==2.2.3",
    "numpy==2.1.3",
    "openpyxl==3.1.5",
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    "python-dotenv==1.0.1",
    "python-dateutil==2.9.0",
    "aiofiles==24.1.0",
    "pypdf==5.1.0",
    "pillow==11.0.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "websockets>=15.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.uv]
# dev-dependencies = []

[project.scripts]
run-api = "app.main:main"
